+++ jmetal/runner/singleobjective/nonElitistEvolutionStrategyRunner.py
@@ -1,4 +1,4 @@
-from jmetal.algorithm.singleobjective.evolutionaryalgorithm.evolutionaryalgorithm import NonElitistEvolutionStrategy
+from jmetal.algorithm.singleobjective.evolutionaryalgorithm import NonElitistEvolutionStrategy
 from jmetal.core.solution import BinarySolution, FloatSolution
 from jmetal.operator.mutation import Polynomial, BitFlip
 from jmetal.problem.singleobjectiveproblem import OneMax, Sphere
